duzelis

+++ BlaBlaCarAz.BLL/DomainModel/Entities/Book.cs
@@ -12,7 +12,7 @@ namespace BlaBlaCarAz.BLL.DomainModel.Entities
         public long AppUserId { get; set; }
         public virtual AppUser AppUser { get; set; }
         public virtual Ride Ride { get; set; }
-        public int RideId { get; set; }
+        public long RideId { get; set; }
         public int LoadLimits { get; set; }
         public bool IsConfirmed { get; set; }
